+++ lib/generators/disco_app/monitorify/monitorify_generator.rb
@@ -1,14 +1,15 @@
 module DiscoApp
   module Generators
-    class RollbarifyGenerator < Rails::Generators::Base
+    class MonitorifyGenerator < Rails::Generators::Base
 
       source_root File.expand_path('../templates', __FILE__)
 
-      # Install the Rollbar and OJ gems.
+      # Install the Rollbar, OJ and New Relic gems.
       def install_gems
         # Add gem to Gemfile
         gem 'rollbar', '~> 2.7.1'
         gem 'oj', '~> 2.14.3'
+        gem 'newrelic_rpm', '~> 3.14.1.311'
 
         # Install gem.
         Bundler.with_clean_env do
@@ -16,9 +17,10 @@ module DiscoApp
         end
       end
 
-      # Copy initializer.
+      # Copy Rollbar initializer and New Relic config file.
       def configure
         copy_file 'initializers/rollbar.rb', 'config/initializers/rollbar.rb'
+        copy_file 'config/newrelic.yml', 'config/newrelic.yml'
       end
 
     end